refactor: organize code into modules
