+++ src/ws.rs
@@ -1,0 +1,53 @@
+use std::sync::Arc;
+
+use axum::{
+    extract::{ws::{Message, WebSocket}, WebSocketUpgrade},
+    response::IntoResponse,
+    Extension,
+};
+use futures::{SinkExt, StreamExt};
+use tokio::sync::{watch, Mutex};
+
+pub struct State {
+    pub clients_count: usize,
+    pub rx: watch::Receiver<Message>,
+}
+
+pub async fn websocket_handler(
+    ws: WebSocketUpgrade,
+    Extension(state): Extension<Arc<Mutex<State>>>,
+) -> impl IntoResponse {
+    ws.on_upgrade(|socket| websocket(socket, state))
+}
+
+async fn websocket(stream: WebSocket, state: Arc<Mutex<State>>) {
+    let (mut sender, mut receiver) = stream.split();
+
+    let mut rx = {
+        let mut state = state.lock().await;
+        state.clients_count += 1;
+        state.rx.clone()
+    };
+
+    let mut send_task = tokio::spawn(async move {
+        while let Ok(()) = rx.changed().await {
+            let msg = rx.borrow().clone();
+
+            if sender.send(msg).await.is_err() {
+                break;
+            }
+        }
+    });
+
+    let mut recv_task = tokio::spawn(async move {
+        while let Some(Ok(_)) = receiver.next().await {}
+    });
+
+    tokio::select! {
+        _ = (&mut send_task) => recv_task.abort(),
+        _ = (&mut recv_task) => send_task.abort(),
+    };
+
+    state.lock().await.clients_count -= 1;
+}
+
